Fix the remaining MSVC /W4 diagnostics in the library
Write UTF-8 byte constants as character escapes instead of truncating
casts (C4310), turn the yyjson read-flag alias into a macro because
yyjson's flags are static const objects rather than constant expressions
(C2099), cast memcpy's pointer-to-const destination to void * (C4090),
and cast the size_t operation counts to the uint32_t parameter (C4267).

+++ src/type/builtin_bitwise.c
@@ -234,5 +234,5 @@ nmo_status_t nmo_register_bitwise_operations(
     };
 
     const size_t count = sizeof(operations) / sizeof(operations[0]);
-    return nmo_operation_registry_register_bulk(operation_registry, operations, count, type_registry, NULL);
+    return nmo_operation_registry_register_bulk(operation_registry, operations, (uint32_t)count, type_registry, NULL);
 }
